+++ BlogposterCMS/app.js
@@ -290,6 +290,7 @@ function getModuleTokenForDbManager() {
     { name:'translationManager',  path:'mother/modules/translationManager',  extra:{} },
     { name:'plainSpace',          path:'mother/modules/plainSpace',          extra:{ app } },
     { name:'themeManager',        path:'mother/modules/themeManager',        extra:{} },
+    { name:'fontsManager',        path:'mother/modules/fontsManager',        extra:{} },
     { name:'importer',            path:'mother/modules/importer',            extra:{} }
   ];